+++ src/cc1/str.h
@@ -1,7 +1,7 @@
 #ifndef STR_H
 #define STR_H
 
-void escape_string(char *str, int *len);
+void escape_string(char *str, size_t *len);
 
 int literal_print(FILE *f, const char *s, int len);
